src/main/java/com/project/WorldFirst/model/User.java
package com.project.WorldFirst.model;

import com.google.common.base.Objects;

public class User {
	
	public int userId = 0;
	public String firstName = null;
	public String surname = null;
	public String email = null;
	
	/**
	 * @return the userId
	 */
	public int getUserId() {
		return userId;
	}
	/**
	 * @return the firstName
	 */
	public String getFirstName() {
		return firstName;
	}
	/**
	 * @return the surname
	 */
	public String getSurname() {
		return surname;
	}
	/**
	 * @return the email
	 */
	public String getEmail() {
		return email;
	}
	/**
	 * @param userId the userId to set
	 */
	public void setUserId(int userId) {
		this.userId = userId;
	}
	/**
	 * @param firstName the firstName to set
	 */
	public void setFirstName(String firstName) {
		this.firstName = firstName;
	}
	/**
	 * @param surname the surname to set
	 */
	public void setSurname(String surname) {
		this.surname = surname;
	}
	/**
	 * @param email the email to set
	 */
	public void setEmail(String email) {
		this.email = email;
	}
	
	@Override
    public boolean equals(Object o) {
        if (this == o) {
            return true;
        }
        if (o == null || getClass() != o.getClass()) {
            return false;
        }
        User user = (User) o;
        return userId == user.getUserId() &&
        		firstName.equals(user.getFirstName()) &&
        		surname.equals(user.getSurname()) &&
        		email.equals(user.getEmail());
    }

	@Override
    public int hashCode() {
        return Objects.hashCode(userId, firstName, surname, email);
    }

}
